requirements.txt
joblib==1.2.0
pandas==1.4.4
Pillow==9.4.0
scikit_learn==1.2.2
streamlit==1.14.0
streamlit_option_menu==0.3.2
